requirements.txt
streamlit==1.32.0
streamlit-folium==0.18.0
folium==0.15.1
pandas==2.2.1
boto3==1.34.51
geopy==2.4.1
python-dotenv==1.0.1
branca==0.7.1
geopandas==0.14.3
requests==2.31.0
shapely==2.0.3
numpy==1.26.4
faker==24.3.0
pytest==8.1.1
pytest-cov==5.0.0
black==24.3.0
flake8==7.0.0
pyproj==3.6.1
fiona==1.9.6
